Rename consumers -> consumer
